enter int

+++ task023/Program.cs
@@ -2,8 +2,8 @@
 //и выдаёт таблицу кубов чисел от 1 до N.
 Console.Clear();
 Console.Write("Введите число : ");
-double num=double.Parse(Console.ReadLine());
-double num2;
+int num=int.Parse(Console.ReadLine());
+int num2;
 for (int i=1;i<=num;i++)
 {
     Console.Write($"{Math.Pow(i,3)} ; ");
